quickreserve/backend_py/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==6.4.0
python-socketio==5.13.0
celery==5.5.3
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
geoalchemy2==0.18.0
